--- garage/torch/core/__init__.py
@@ -1,6 +1,0 @@
-"""
-General classes, functions, utilities that are used throughout garage.torch.
-"""
-from garage.torch.core.logging import logger
-
-__all__ = ['logger']
